feat, topmost window, exclude from capture
