+++ src/lang/ua.rs
@@ -569,6 +569,7 @@ pub static ref T: std::collections::HashMap<&'static str, &'static str> =
         ("Enable blocking user input", "Блокувати введення для користувача"),
         ("id_input_tip", "Ви можете ввести ID, безпосередню IP, або ж домен з портом (<домен>:<порт>).\nЯкщо ви хочете отримати доступ до пристрою на іншому сервері, будь ласка, додайте адресу сервера (<id>@<адреса_сервера>?key=<значення_ключа>), наприклад,\n9123456234@192.168.16.1:21117?key=5Qbwsde3unUcJBtrx9ZkvUmwFNoExHzpryHuPUdqlWM=.\nЯкщо ви хочете отримати доступ до пристрою на публічному сервері, будь ласка, введіть \"<id>@public\", ключ для публічного сервера не потрібен."),
         ("privacy_mode_impl_mag_tip", "Режим 1 (застарілий)"),
+        ("privacy_mode_impl_exclude_from_capture_tip", ""),
         ("privacy_mode_impl_virtual_display_tip", "Режим 2 (рекомендований)"),
         ("Enter privacy mode", "Увійти в режим конфіденційності"),
         ("Exit privacy mode", "Вийти з режиму конфіденційності"),
