+++ src/privacy_mode/win_exclude_from_capture.rs
@@ -1,0 +1,3 @@
+pub use super::win_topmost_window::PrivacyModeImpl;
+
+pub(super) const PRIVACY_MODE_IMPL: &str = "privacy_mode_impl_exclude_from_capture";
